Solved migrations Errors and updated Frontend

+++ Day25/PizzaStoreManagmentAPI/Models/Pizza.cs
@@ -9,7 +9,7 @@ namespace PizzaStoreManagmentAPI.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string size { get; set; }
-        public float price { get; set; }
+        public int price { get; set; }
 
         public string Availability { get; set; }
     }
